cleanup model hierarchy 1
