+++ edgy/core/db/models/managers.py
@@ -4,17 +4,17 @@ from edgy.core.db.context_vars import get_tenant, set_tenant
 from edgy.core.db.querysets.base import QuerySet
 
 if TYPE_CHECKING:
-    from edgy.core.db.models.base import EdgyBaseModel
+    from edgy.core.db.models.base import BaseModelType
 
 
 class BaseManager:
     def __init__(
         self,
         *,
-        owner: Optional[Union[Type["EdgyBaseModel"]]] = None,
+        owner: Optional[Union[Type["BaseModelType"]]] = None,
         inherit: bool = True,
         name: str = "",
-        instance: Optional[Union["EdgyBaseModel"]] = None,
+        instance: Optional[Union["BaseModelType"]] = None,
     ):
         self.owner = owner
         self.inherit = inherit
