+++ edgy/core/db/models/mixins/__init__.py
@@ -1,3 +1,5 @@
 from .generics import DeclarativeMixin
+from .reflection import ReflectedModelMixin
+from .row import ModelRowMixin
 
-__all__ = ["DeclarativeMixin"]
+__all__ = ["DeclarativeMixin", "ModelRowMixin", "ReflectedModelMixin"]
